requirements.txt
streamlit>=1.33.0
requests
beautifulsoup4
pdfplumber
pyyaml
google-genai
